requirements.txt
flask==3.1.1
psycopg2==2.9.10
python-dotenv==1.1.0